refactor(tts): share auto mode normalization
